First return 😅

+++ HomeWork/28_ProcessHandling/28_ProcessHandling/Core/Implementations/LinqSumProcessor.cs
@@ -4,10 +4,10 @@ namespace _28_ProcessHandling.Core.Implementations;
 
 public class LinqSumProcessor : ISumProcessing
 {
-    public long ProcessSum(int[] numbers)
+    public long ProcessSum(long[] numbers)
     {
-        return numbers.AsParallel().Sum(number => (long)number);
+        return numbers.AsParallel().Sum(number => number);
     }
-    
+
     public string Title => "✨  LINQ: AsParallel";
 }
