+++ HomeWork/28_ProcessHandling/28_ProcessHandling/Core/Interfaces/ISumProcessing.cs
@@ -2,7 +2,7 @@ namespace _28_ProcessHandling.Core.Interfaces;
 
 public interface ISumProcessing
 {
-    long ProcessSum(int[] numbers);
+    long ProcessSum(long[] numbers);
     
     string Title { get; }
 }
